add file for constants - move alert constants to file - remove link from warning on save
