+++ src/datadoc/constants.py
@@ -1,0 +1,6 @@
+"""Repository for constant values in Datadoc."""
+
+MISSING_METADATA_WARNING = "Advarsel - obligatorisk metadata mangler"
+CHECK_OBLIGATORY_METADATA_DATASET_MESSAGE = "Følgende datasett felt har ikke verdi:"
+CHECK_OBLIGATORY_METADATA_VARIABLES_MESSAGE = "Følgende variabler har felt uten verdi:"
+DAPLA_MANUAL_TEXT = "Dapla manual navnestandard"
