+++ src/datadoc/frontend/callbacks/utils.py
@@ -33,8 +33,6 @@ MetadataInputTypes: TypeAlias = (
     str | list[str] | int | float | bool | datetime.date | None
 )
 
-MISSING_METADATA_WARNING = "Advarsel - obligatorisk metadata ikke utfylt"
-
 
 def get_language_strings_enum(
     enum: Enum | type[enums.LanguageStringsEnum],
